Extract all scaffolds within a selection polygon

Finds all scaffolds that fall into a selection polygon highlighted in a plot generated by mmplot.

mmextract(mm, selection, min_length = 0, inverse = FALSE)

Arguments

mm
(required) A dataframe loaded with mmload.

selection
(required) A 2-column dataframe with the x and y coordinates of a selection of points in an mmplot. The column names of the provided dataframe must match column names in mm.

min_length
Filter all scaffolds with a length at or below this value before the extraction. (Default: 0)

inverse
(Logical) If TRUE, then the scaffolds within the selection are instead removed. (Default: FALSE)

Value

A dataframe (tibble) compatible with other mmgenome2 functions.

Examples

library(mmgenome2)
data(mmgenome2)
mmgenome2

#> # A tibble: 97,285 x 10
#>    scaffold length    gc cov_C13.11.14 cov_C13.11.25 cov_C13.12.03
#>    <chr>     <dbl> <dbl>         <dbl>         <dbl>         <dbl>
#>  1 1         8264.  57.8       1.44             53.6         0.   
#>  2 2         1027.  57.0       0.625            24.2         0.   
#>  3 3         1665.  55.9      13.5             434.          0.166
#>  4 4         9056.  35.9       0.0100           23.4         0.   
#>  5 5         3343.  64.0       3.20             16.4         0.   
#>  6 6        98207.  39.1       0.00966          24.5         3.29 
#>  7 7         6480.  63.0       2.61             19.2         1.46 
#>  8 8        15790.  61.7       2.78             21.2         1.62 
#>  9 9         1403.  70.4      85.1             192.          0.   
#> 10 10        2018.  70.2      50.3             101.          0.   
#> # ... with 97,275 more rows, and 4 more variables: cov_C14.01.09 <dbl>,
#> #   geneID <chr>, phylum <fct>, rRNA16S <fct>

selection <- data.frame(cov_C14.01.09 = c(24.852, 32.545, 53.062, 38.52),
                        cov_C13.12.03 = c(7.676, 5.165, 6.386, 10.933))
mmgenome2_extraction <- mmextract(mmgenome2,
                                  min_length = 3000,
                                  selection = selection,
                                  inverse = FALSE)

#> 83 scaffolds (or 2.58% of the scaffolds in mm, weighted by length) remain after 97202 of 97285 scaffolds have been filtered.

mmgenome2_extraction

#> # A tibble: 83 x 10
#>    scaffold  length    gc cov_C13.11.14 cov_C13.11.25 cov_C13.12.03
#>    <chr>      <dbl> <dbl>         <dbl>         <dbl>         <dbl>
#>  1 180       22093.  74.0         0.298          12.0          7.34
#>  2 227      190826.  65.4         0.348          13.0          6.91
#>  3 253      296715.  70.2         1.06           15.2          7.00
#>  4 315      218862.  71.7         0.366          12.6          7.17
#>  5 386      164942.  68.2         0.359          12.3          6.73
#>  6 396        7448.  72.6         0.329          13.0          7.59
#>  7 423      100199.  68.7         1.40           15.3          6.71
#>  8 431      446220.  72.4         0.409          12.6          6.80
#>  9 460      420964.  71.1         0.421          13.4          6.97
#> 10 484      454223.  72.6         0.379          12.6          7.09
#> # ... with 73 more rows, and 4 more variables: cov_C14.01.09 <dbl>,
#> #   geneID <chr>, phylum <fct>, rRNA16S <fct>