Network plot of mate-pair or paired-end connected scaffold reads

Network plot of mate-pair or paired-end connected scaffold reads

mmnetwork(mm, network, min_connections = 2, color_by = NULL,
  color_scale_log10 = FALSE, locator = FALSE, selection = NULL,
  highlight_labels = NULL, highlight_color = "darkred", links_scale = 1,
  scaffold_labels = FALSE, print_nolinks = FALSE, seed = .mm_Random.seed)

Arguments

mm
(required) A dataframe loaded with mmload.

network
(required) Paired-end or mate-pair connections between scaffolds in long format. First and second column should contain all connected scaffold pairs and the third column the number of connections.

min_connections
Filter all scaffold pairs with equal to or less than this number of connections. (Default: 2)

color_by
Color the scaffolds by a variable in mm. (Default: NULL)

color_scale_log10
(Logical) Log10-scale the colors of the variable defined by color_by. (Default: FALSE)

locator
(Logical) When enabled (set to TRUE), left-clicks in the plot are captured and the exact x/y-coordinates of the mouse clicks are captured and returned. These coordinates can be used to highlight a selection of scaffolds in the plot, and to extract all scaffolds within the selection. (Default: FALSE)

selection
A 2-column dataframe with the x and y coordinates of points with which to draw a polygon onto the plot to highlight a selected region. Can be obtained by using the locator feature (by locator = TRUE). (Default: NULL)

highlight_labels
A dataframe or vector of scaffold names whose labels to highlight in the plot (colored by highlight_color). (Default: NULL)

highlight_color
The color with which to highlight the scaffold labels defined by highlight_labels. (Default: "darkred")

links_scale
A factor to scale the sizes of the links plotted between scaffolds. (Default: 1)

scaffold_labels
Add labels with the scaffold names of all scaffolds.

print_nolinks
(Logical) Print the names of all scaffolds with no links to other scaffolds to the console. (Default: FALSE)

seed
Network plots are based on Random Number Generation, and this is used to set a specific random seed (with set.seed) allowing for reproducible network plots. Set to NULL to generate a unique network plot with each run of mmnetwork, the random seed in R is always available in the hidden object .Random.seed. (Default: .mm_Random.seed)

Using the locator with mmnetwork

Unlike a plot generated by mmplot, the coordinates of points in each dimension in a mmnetwork plot is not contained within the provided dataframe mm itself. As such mmnetwork does not have a corresponding extract function like mmextract is for mmplot. Instead, a subset of mm containing all scaffolds within the selection polygon is available in the returned ggplot object with plot$data_in_selection.