Visualise metagenomes in various ways

mmplot(mm, x, y, min_length = 0, color_by = NULL, locator = FALSE,
  selection = NULL, x_scale = NULL, x_limits = NULL, y_scale = NULL,
  y_limits = NULL, alpha = 0.1, scaffold_labels = FALSE,
  fixed_size = NULL, size_scale = 1, shared_genes = FALSE)

Arguments

mm
(required) A dataframe loaded with mmload.

x
(required) The variable from mm to plot on the first axis.

y
(required) The variable from mm to plot on the second axis.

min_length
Remove scaffolds with a length at or below this value before plotting. (Default: 0)

color_by
Color the scaffolds by a variable in mm. (Default: NULL)

locator
(Logical) When enabled (set to TRUE), left-clicks in the plot are captured and the exact x/y-coordinates of the mouse clicks are returned. These coordinates can be used to highlight a selection of scaffolds in the plot, and also be used with mmextract to extract all scaffolds within the selection from the data. (Default: FALSE)

selection
A 2-column dataframe with the x and y coordinates of points with which to draw a polygon onto the plot to highlight a selected region. Can be obtained by using the locator feature (by locator = TRUE). (Default: NULL)

x_scale
Log10-scale ("log10") or square-root scale "sqrt" the x axis. (Default: NULL)

x_limits
Axis limits of the x axis. (Default: NULL)

y_scale
Log10-scale ("log10") or square-root scale "sqrt" the y axis. (Default: NULL)

y_limits
Axis limits of the y axis. (Default: NULL)

alpha
The transparancy of the scaffold points, where 0 is invisible and 1 is opaque. (Default: 0.1)

scaffold_labels
Add labels of a selection of scaffolds by providing either a character vector of scaffold names, or a dataframe with scaffold names in the first column. If set to TRUE then all scaffolds will be labelled. (Default: FALSE)

fixed_size
A fixed size for all scaffolds if set. If NULL then the scaffolds are scaled by length. (Default: NULL)

size_scale
A factor to scale the sizes of the scaffolds plotted. Only applies when fixed_size is set to NULL and the scaffolds are scaled by length. (Default: 1)

shared_genes
(Logical) If TRUE, lines will be drawn between scaffolds with any shared gene(s). (Default: FALSE)