Interactively define a selection subspace on mmplot or mmplot_network

Interactively define a selection subspace on mmplot and mmplot_network. The selection can be used to extract subsets of mmgenome data objects using mmextract. The function works on generic ggplot2 objects,

mmplot_locator(p, sig)

Arguments

p
(required) A mmplot or mmplot_network object.

sig
The number of significant digits (default: 3).

trans
Specify x and y-axis transformation used in mmplot. The format is trans=c(x-axis, y-axis) i.e. trans=c("", "log10"). The following transformations are supported: No transformation (default) = "". Log10 = "log10". Square root transformation = "sqrt".

Value

X and Y coordinates marked in the plot or scaffold names.

Examples

data(rocco)

p <- mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T, log.y = T, color = "pps_phylum", minlength = 5000)

#> Error in mmplot(data = d, x = "C13.12.03", y = "C14.01.09", log.x = T,     log.y = T, color = "pps_phylum", minlength = 5000): unused arguments (data = d, log.x = T, log.y = T, minlength = 5000)

#sel <- mmplot_locator(p, trans=c("log10", "log10"))
sel <- data.frame(C13.12.03  =  c(1.39, 2.07, 16.8, 19.4, 7.72, 1.76),
                  C14.01.09  =  c(29.4, 67.6, 85.9, 43.6, 16.7, 14.9))
mmplot_selection(p, sel)

#> Error in mmplot_selection(p, sel): object 'p' not found

dA <- mmextract(d, sel)

#> Error in unlist(data$scaffolds[xname]): object 'd' not found

pn <- mmplot_network(data = dA, network = pe, color = "essential", nconnections = 10)

#> Error in subset(network, network$scaffold1 %in% data$scaffolds$scaffold &     network$scaffold2 %in% data$scaffolds$scaffold & network$connections >=     nconnections): object 'pe' not found

#sel <- mmplot_locator(pn)
sel <- c("3951", "6944")
mmplot_network(data = dA, network = pe, color = "essential", nconnections = 10, highlight= sel)

#> Error in subset(network, network$scaffold1 %in% data$scaffolds$scaffold &     network$scaffold2 %in% data$scaffolds$scaffold & network$connections >=     nconnections): object 'pe' not found