Compare essential genes in a bin to a group of references

Compare essential genes in an extracted genome bin to a group of references

mmref(data, tax.level, tax.compare)

Arguments

data
The dataframe containing the genome bin of interest (optional).

tax.level
What level to compare on. Supports Kingdom, Phylum, Class, Genus, Species (default: Phylum)

tax.compare
A vector with names on what should be used for comparison (e.g. c("Proteobacteria", "Actinobacteria")) or "all".

tax.aggregate
Summarise the data to a fixed taxonomic rank. Supports Kingdom, Phylum, Class, Genus, Species (default: tax.level).

summarise
Report "mean" or "median" HMM counts (default: mean).

output
Either complete or plot (default: plot)

plot.display
Show all HMMs even if they are not contained in the selected references (default: T).

plot.sort
Sort the plot (default: T)

Value

An overview of essential genes in the bin and selected reference genomes.

Examples

mmref(tax.level = "Phylum", tax.compare = c("Proteobacteria", "Actinobacteria"))

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

res <- mmref(tax.level = "Phylum", tax.compare = c("Proteobacteria", "Actinobacteria"), output = "complete")

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

res$table

#> Error in eval(expr, envir, enclos): object 'res' not found

mmref(tax.level = "Phylum", tax.compare = "all")

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class")

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class", summarise = "median")

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

mmref(tax.level = "Class", tax.compare = "Betaproteobacteria", tax.aggregate = "Strain", summarise = "median")

#> Error in ddply(em, c(tax.aggregate, "HMM"), summarize, median = median(Count),     mean = mean(Count)): could not find function "ddply"

mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class", summarise = "median", data = dA)

#> Error in mmref(tax.level = "Phylum", tax.compare = "Proteobacteria", tax.aggregate = "Class",     summarise = "median", data = dA): object 'dA' not found