Export DNA sequences of selected scaffolds

Writes the DNA sequences of a given set of scaffolds from the assembly to a FASTA file.

mmexport(scaffolds, assembly = get("assembly", envir = .GlobalEnv),
  file = "exported_assembly.fa")

Arguments

scaffolds
(required) Either a vector of scaffold names or a dataframe loaded with mmload containing the scaffolds to extract from the assembly.

assembly
(required) The assembly to subset and export. Must be an object of class "DNAStringSet" as loaded with readDNAStringSet. By default any object named "assembly" in the global environment will be used.

file
The file path and file name of the file to write the sequences to. (Default: "exported_assembly.fa")

Examples

# NOT RUN {
library(mmgenome2)
data(mmgenome2)
mmgenome2
selection <- data.frame(cov_C14.01.09 = c(24.852, 32.545, 53.062, 38.52),
                        cov_C13.12.03 = c(7.676, 5.165, 6.386, 10.933))
mmgenome2_extraction <- mmextract(mmgenome2,
                                  min_length = 3000,
                                  selection = selection,
                                  inverse = FALSE)
mmgenome2_extraction
mmexport(mmgenome2_extraction,
         assembly = assembly,
         file = "exported_assembly.fa")
# }