Export DNA sequences of selected scaffolds

Writes the DNA sequences of a given set of scaffolds from the assembly to a FASTA file.

mmexport(scaffolds, assembly = get("assembly", envir = .GlobalEnv),
  file = "exported_assembly.fa")

Arguments

scaffolds
(required) Either a vector of scaffold names or a dataframe loaded with mmload containing the scaffolds to extract from the assembly.

assembly
(required) The assembly to subset and export. Must be an object of class "DNAStringSet" as loaded with readDNAStringSet. By default any object named "assembly" in the global environment will be used.

file
The file path and file name of the file to write the sequences to. (Default: "exported_assembly.fa")

Examples

library(mmgenome2)
data(mmgenome2)
mmgenome2

#> [38;5;246m# A tibble: 97,285 x 13[39m
#>    scaffold length    gc cov_C13.11.14 cov_C13.11.25 cov_C13.12.03 cov_C14.01.09
#>    [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 1          [4m8[24m264  57.8       1.44             53.6         0             0.066
#> [38;5;250m 2[39m 2          [4m1[24m027  57.0       0.625            24.2         0             0    
#> [38;5;250m 3[39m 3          [4m1[24m665  55.9      13.5             434.          0.166         0.177
#> [38;5;250m 4[39m 4          [4m9[24m056  35.9       0.01             23.4         0             0    
#> [38;5;250m 5[39m 5          [4m3[24m343  64.0       3.20             16.4         0             0    
#> [38;5;250m 6[39m 6         [4m9[24m[4m8[24m207  39.1       0.009[4m6[24m[4m6[24m          24.5         3.29          9.85 
#> [38;5;250m 7[39m 7          [4m6[24m480  63.0       2.61             19.2         1.46         12.3  
#> [38;5;250m 8[39m 8         [4m1[24m[4m5[24m790  61.7       2.78             21.2         1.62         10.3  
#> [38;5;250m 9[39m 9          [4m1[24m403  70.4      85.1             192.          0             0    
#> [38;5;250m10[39m 10         [4m2[24m018  70.2      50.3             101.          0             0    
#> [38;5;246m# ... with 97,275 more rows, and 6 more variables: PC1 [3m[38;5;246m<dbl>[38;5;246m[23m, PC2 [3m[38;5;246m<dbl>[38;5;246m[23m,
#> #   PC3 [3m[38;5;246m<dbl>[38;5;246m[23m, geneID [3m[38;5;246m<chr>[38;5;246m[23m, taxonomy [3m[38;5;246m<fct>[38;5;246m[23m, rRNA16S [3m[38;5;246m<fct>[38;5;246m[23m[39m

selection <- data.frame(
  cov_C13.11.25 = c(7.2, 16.2, 25.2, 23.3, 10.1),
  cov_C14.01.09 = c(47, 77, 52.8, 29.5, 22.1)
)
mmgenome2_extraction <- mmextract(mmgenome2,
  min_length = 3000,
  selection = selection,
  inverse = FALSE
)

#> 47 scaffolds (or 1.33% of the scaffolds in mm, weighted by length) remain after 97238 of 97285 scaffolds have been filtered.

mmgenome2_extraction

#> [38;5;246m# A tibble: 47 x 13[39m
#>    scaffold length    gc cov_C13.11.14 cov_C13.11.25 cov_C13.12.03 cov_C14.01.09
#>    [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m 180       [4m2[24m[4m2[24m093  74.0         0.298          12.0          7.34          37.9
#> [38;5;250m 2[39m 227      [4m1[24m[4m9[24m[4m0[24m826  65.4         0.348          13.0          6.91          38.5
#> [38;5;250m 3[39m 253      [4m2[24m[4m9[24m[4m6[24m715  70.2         1.06           15.2          7.00          41.1
#> [38;5;250m 4[39m 315      [4m2[24m[4m1[24m[4m8[24m862  71.7         0.366          12.6          7.17          38.8
#> [38;5;250m 5[39m 343       [4m2[24m[4m7[24m203  57.6         0.235          10.7          5.07          32.0
#> [38;5;250m 6[39m 381        [4m6[24m128  67.2         8.11           22.6          5.89          48.7
#> [38;5;250m 7[39m 386      [4m1[24m[4m6[24m[4m4[24m942  68.2         0.359          12.3          6.73          37.2
#> [38;5;250m 8[39m 396        [4m7[24m448  72.6         0.329          13.0          7.59          41.0
#> [38;5;250m 9[39m 423      [4m1[24m[4m0[24m[4m0[24m199  68.7         1.40           15.3          6.71          40.0
#> [38;5;250m10[39m 431      [4m4[24m[4m4[24m[4m6[24m220  72.4         0.409          12.6          6.80          37.9
#> [38;5;246m# ... with 37 more rows, and 6 more variables: PC1 [3m[38;5;246m<dbl>[38;5;246m[23m, PC2 [3m[38;5;246m<dbl>[38;5;246m[23m, PC3 [3m[38;5;246m<dbl>[38;5;246m[23m,
#> #   geneID [3m[38;5;246m<chr>[38;5;246m[23m, taxonomy [3m[38;5;246m<fct>[38;5;246m[23m, rRNA16S [3m[38;5;246m<fct>[38;5;246m[23m[39m
# NOT RUN {
mmexport(mmgenome2_extraction,
  assembly = assembly,
  file = "bins/exported_assembly.fa"
)
# }