Export DNA sequences of selected scaffolds

Writes the DNA sequences of a set of scaffolds from the assembly to a FASTA file.

mmexport(scaffolds, assembly = assembly, file = "exported_assembly.fa")

Arguments

scaffolds
(required) Either a vector of scaffold names or a dataframe loaded with mmload containing the scaffolds to extract from the assembly.

assembly
(required) The assembly to subset and export. Must be an object of class "DNAStringSet" as loaded with readDNAStringSet.

file
The file path and file name of the file to write the sequences to. (Default: "exported_assembly.fa")