Print basic statistics of metagenome data

Calculates some basic statistics like the N50 length, minimum-, mean-, and maximum scaffold lengths, mean GC content, information about essential genes, and more...

mmstats(mm)

Arguments

mm
(required) A dataframe loaded with mmload.

Value

A dataframe with the calculated stats.

Details

The returned stats are calculated as follows:

Scaffolds: The number of different scaffolds in the assembly.

N50: The shortest sequence length at 50

Length.total: The total size of the assembly.

Length.max: The size of the largest scaffold in the assembly.

Length.mean: The average scaffold size in the assembly.

Length.min: The size of the smallest scaffold in the assembly.

weighted_GC_mean: The average GC content in the assembly, weighted by scaffold sizes.

cov_*: The average coverage of each coverage variable in mm, weighted by scaffold sizes. (Only columns starting with "cov_" will be shown)

Ess.genes.total: The total number of essential genes, if any have been loaded.

Ess.genes.unique: The number of unique essential genes, if any have been loaded.

Examples

library(mmgenome2)
data(mmgenome2)
mmgenome2

#> # A tibble: 97,285 x 13
#>    scaffold length    gc cov_C13.11.14 cov_C13.11.25 cov_C13.12.03
#>    <chr>     <dbl> <dbl>         <dbl>         <dbl>         <dbl>
#>  1 1          8264  57.8       1.44             53.6         0    
#>  2 2          1027  57.0       0.625            24.2         0    
#>  3 3          1665  55.9      13.5             434.          0.166
#>  4 4          9056  35.9       0.01             23.4         0    
#>  5 5          3343  64.0       3.20             16.4         0    
#>  6 6         98207  39.1       0.00966          24.5         3.29 
#>  7 7          6480  63.0       2.61             19.2         1.46 
#>  8 8         15790  61.7       2.78             21.2         1.62 
#>  9 9          1403  70.4      85.1             192.          0    
#> 10 10         2018  70.2      50.3             101.          0    
#> # ... with 97,275 more rows, and 7 more variables: cov_C14.01.09 <dbl>,
#> #   PC1 <dbl>, PC2 <dbl>, PC3 <dbl>, geneID <chr>, taxonomy <fct>,
#> #   rRNA16S <fct>

mmstats(mmgenome2)

#>                  General stats
#> Scaffolds                97285
#> N50                       5291
#> Length.total         331908376
#> Length.max             1446979
#> Length.mean            3411.71
#> Length.min                1000
#> weighted_GC_mean         52.22
#> cov_C13.11.14             2.62
#> cov_C13.11.25            16.06
#> cov_C13.12.03             2.91
#> cov_C14.01.09            11.58
#> Ess.genes.total           7138
#> Ess.genes.unique           109